+++ runner.go
@@ -122,7 +122,26 @@ func RunSuite[Suite suite[T], T CommonT](
 
 	r := newRunner[Suite]()
 
-	return r.runSuite(testingT, suite, options...)
+	return r.runSuite(testingT, suite, nil, options...)
+}
+
+// RunSubSuite runs a sub-suite.
+//
+// This is similar to [RunSuite] but designed to be called from other suites.
+//
+// RunSubSuite reports whether all sub-suite tests succeeded.
+//
+// NOTE: this function may cause infinite loop if called within the same suite as passed to it.
+func RunSubSuite[Suite suite[Sub], Parent, Sub CommonT](
+	t Parent,
+	suite Suite,
+	options ...testoplugin.Option,
+) bool {
+	t.Helper()
+
+	r := newRunner[Suite]()
+
+	return r.runSuite(t.unwrap().testingT, suite, &t.unwrap().reflection.Suite, options...)
 }
 
 // Run runs f as a subtest of t called name. It runs f in a separate goroutine
@@ -161,6 +180,7 @@ func Run[T CommonT](
 					RawBaseName: name,
 					Level:       t.level(),
 					IsSubtest:   true,
+					FuncPC:      reflect.ValueOf(f).Pointer(),
 				}
 			},
 			options...,
@@ -222,6 +242,7 @@ func (r *runner[Suite, T]) collectTests(
 func (r *runner[Suite, T]) runSuite(
 	testingT TestingT,
 	suite Suite,
+	parentSuite *testoreflect.SuiteInfo,
 	options ...testoplugin.Option,
 ) bool {
 	testingT.Helper()
@@ -233,6 +254,7 @@ func (r *runner[Suite, T]) runSuite(
 	tests := r.collectTests(testingT, caller, suite)
 
 	suiteInfo := testoreflect.SuiteInfo{
+		Parent:   parentSuite,
 		Name:     r.suiteName,
 		Caller:   testingT.Name(),
 		TestingT: testingT,
@@ -290,6 +312,7 @@ func (r *runner[Suite, T]) runSuiteTests(t T, s Suite, tests suiteTests[Suite, T
 	s.BeforeAll(t)
 
 	suiteInfo := testoreflect.SuiteInfo{
+		Parent:   t.unwrap().reflection.Suite.Parent,
 		Name:     t.unwrap().reflection.Suite.Name,
 		Caller:   t.unwrap().reflection.Suite.Caller,
 		TestingT: t.unwrap().reflection.Suite.TestingT,