+++ testoreflect/reflect.go
@@ -200,6 +200,10 @@ func (ParametrizedTestInfo) isTestInfo() {}
 
 // SuiteInfo is the information about suite.
 type SuiteInfo struct {
+	// Parent refers to the parent suite info.
+	// Non-nil value means that current suite is sub-suite.
+	Parent *SuiteInfo
+
 	// Name of this suite.
 	Name string
 
@@ -276,7 +280,7 @@ type RegularTestInfo struct {
 
 	// FuncPC is the program counter (PC) of this test function.
 	//
-	// NOTE: it may be empty (zero) in some cases, e.g. hooks and sub-tests.
+	// NOTE: it may be empty (zero) in some cases, e.g. hooks.
 	// Use with caution.
 	FuncPC uintptr
 }